backend/requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-CORS==4.0.0
PyJWT==2.8.1
requests==2.31.0
python-dotenv==1.0.0
PyMySQL==1.1.0
redis==5.0.0
cryptography==41.0.1
Werkzeug==2.3.0

# 并发测试依赖
locust==2.15.1

